Refactoring project, logic saldo added, LogistaController, UserComumController, TransferController removed, OutDTOs and DTOs added, services refactored, Usuario entity changed, more business rules added, some bug fix

+++ src/main/java/dev/marcus/picPaySimplificado/domain/entities/transaction/DTOs/impl/TransactionOutDTOImpl.java
@@ -1,6 +1,5 @@
 package dev.marcus.picPaySimplificado.domain.entities.transaction.DTOs.impl;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import dev.marcus.picPaySimplificado.domain.entities.transaction.Transaction;
@@ -12,14 +11,12 @@ public record TransactionOutDTOImpl(
     UUID id,
     TypeTransaction typeTransaction,
     UsuarioOutDTO usuarioData,
-    BigDecimal valor
+    float valor
 
 ) implements TransactionOutDTO{
-
     public TransactionOutDTOImpl(
         UsuarioOutDTO usuarioData,
         Transaction transaction
-
     ){
         this(
             transaction.getId(),
